modules/edit_pdf.py
from flask import Blueprint, request, jsonify, current_app
import os
from werkzeug.utils import secure_filename
from PyPDF2 import PdfReader, PdfWriter
import uuid
import time
import logging
from reportlab.pdfgen import canvas
from reportlab.lib.pagesizes import letter
from reportlab.lib.colors import Color
from reportlab.pdfbase import pdfmetrics
from reportlab.pdfbase.ttfonts import TTFont
from PIL import Image

# Configure logging
logger = logging.getLogger(__name__)

# Create blueprint
edit_bp = Blueprint('edit', __name__)

def register_routes(app, socketio):
    """Register routes with the Flask app"""
    app.register_blueprint(edit_bp)
    
    # Add routes for PDF editing here
    # This is a placeholder for future implementation of PDF editing features
    # such as adding text, images, signatures, highlighting, etc.
    
    @app.route('/edit', methods=['POST'])
    def edit_route():
        """Placeholder for PDF editing"""
        socketio.emit('status_update', {
            'status': 'error', 
            'message': '⚠️ PDF editing is not yet implemented.',
            'tool': 'edit'
        })
        return jsonify({'status': 'error', 'message': 'PDF editing is not yet implemented.'})

# Functions for PDF editing can be added here in the future
# For example:
# - add_text_to_pdf
# - add_image_to_pdf
# - add_signature_to_pdf
# - highlight_text_in_pdf
# - draw_on_pdf
# etc.