requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
PyPDF2==3.0.1
pdf2image==1.16.3
img2pdf==0.4.4
pdfminer.six==20221105
Pillow==10.0.0
python-docx==0.8.11
tabula-py==2.7.0
pandas==2.0.3
python-pptx==0.6.21
reportlab==3.6.13
pytesseract==0.3.10
python-engineio==4.5.1
python-socketio==5.8.0
# pywin32==306 - Commented out as it's Windows-specific and will cause issues on Vercel
Werkzeug==2.3.7
gunicorn==21.2.0
eventlet==0.33.3
docx2pdf==0.1.8
matplotlib==3.7.2